+++ src/pages/Auth/Join/JoinForm.jsx
@@ -1,0 +1,161 @@
+import { Button, Field, Input } from "@/components";
+import styles from "./JoinForm.module.css";
+import { useNavigate } from "react-router-dom";
+import { PageEndPoints } from "@/constants";
+
+const JoinForm = ({ joinUseForm, onSubmit }) => {
+  const navigate = useNavigate();
+
+  const {
+    register,
+    formState: { errors },
+    handleSubmit,
+  } = joinUseForm;
+
+  return (
+    <form className={styles.form_container} onSubmit={handleSubmit(onSubmit)}>
+      <div className={styles.name}>
+        <Field
+          label="성"
+          error={errors.firstName}
+          isRequire
+          style={{
+            flex: 1,
+          }}
+        >
+          <Input
+            register={register}
+            name={"firstName"}
+            style={{
+              boxSizing: "border-box",
+              width: "100%",
+            }}
+            size="sm"
+          />
+        </Field>
+        <Field
+          label="이름"
+          error={errors.lastName}
+          isRequire
+          style={{
+            flex: 1,
+          }}
+        >
+          <Input
+            register={register}
+            name={"lastName"}
+            style={{
+              boxSizing: "border-box",
+              width: "100%",
+            }}
+            size="sm"
+          />
+        </Field>
+      </div>
+
+      <Field label={"이메일"} error={errors.email} isRequire>
+        <div className={styles.email}>
+          <Input
+            type="text"
+            register={register}
+            name={"email"}
+            style={{
+              boxSizing: "border-box",
+              width: "100%",
+            }}
+            size="sm"
+          />
+          <Button
+            type="button"
+            size="sm"
+            style={{ width: "5.5rem" }}
+            onClick={() => {
+              if (errors.email) {
+                console.log("eamil error");
+              }
+            }}
+          >
+            중복확인
+          </Button>
+        </div>
+      </Field>
+
+      <Field label={"닉네임"} error={errors.nickname} isRequire>
+        <div className={styles.nickname}>
+          <Input
+            type="text"
+            register={register}
+            name={"nickname"}
+            style={{
+              boxSizing: "border-box",
+              width: "100%",
+            }}
+            size="sm"
+          />
+          <Button
+            type="button"
+            size="sm"
+            style={{
+              width: "5.5rem",
+            }}
+          >
+            중복확인
+          </Button>
+        </div>
+      </Field>
+
+      <Field label={"비밀번호"} error={errors.password} isRequire>
+        <Input
+          type="password"
+          register={register}
+          name={"password"}
+          style={{
+            boxSizing: "border-box",
+            width: "100%",
+          }}
+          size="sm"
+        />
+      </Field>
+
+      <Field label={"비밀번호 확인"} error={errors.passwordConfirm} isRequire>
+        <Input
+          type="password"
+          register={register}
+          name={"passwordConfirm"}
+          style={{
+            boxSizing: "border-box",
+            width: "100%",
+          }}
+          size="sm"
+        />
+      </Field>
+
+      <div className={styles.footer}>
+        <Button
+          style={{
+            width: "100%",
+          }}
+          size="lg"
+        >
+          JOIN
+        </Button>
+        <div className={styles.divider} />
+        <div className={styles.login_link_box}>
+          <div>Already have an account?</div>
+          <Button
+            type="button"
+            style={{
+              color: "var(--color-blue-500)",
+            }}
+            variant="none"
+            onClick={() => navigate(PageEndPoints.LOGIN)}
+          >
+            Login
+          </Button>
+        </div>
+      </div>
+    </form>
+  );
+};
+
+export default JoinForm;
