+++ src/pages/Auth/Join/JoinForm.module.css
@@ -1,0 +1,57 @@
+.form_container {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+
+  flex: 1;
+}
+
+.name {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.email {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.nickname {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.divider {
+  width: 100%;
+  border: 0;
+  height: 2px;
+  background-image: linear-gradient(
+    to right,
+    transparent,
+    var(--color-gray-300),
+    transparent
+  );
+  text-align: center;
+}
+
+.divider::after {
+  content: "OR";
+  padding: 0 0.5rem;
+  position: relative;
+  top: -0.4rem;
+  background-color: var(--color-bg-primary);
+  color: var(--color-gray-300);
+  font-weight: var(--font-weight-semibold);
+}
+
+.footer {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
+.login_link_box {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
